+++ pynautobot/core/endpoint.py
@@ -15,7 +15,9 @@ limitations under the License.
 
 This file has been modified by NetworktoCode, LLC.
 """
-from typing import Dict
+
+from typing import List, Dict, Any
+from uuid import UUID
 from pynautobot.core.query import Request, RequestError
 from pynautobot.core.response import Record
 
@@ -309,34 +311,71 @@ class Endpoint(object):
 
         return response_loader(req, self.return_obj, self)
 
-    def update(self, id: str, data: Dict[str, any]):
-        """
-        Update a resource with a dictionary.
+    def update(self, *args, **kwargs):
+        r"""
+        Update a single resource with a dictionary or bulk update a list of objects.
 
-        Accepts the id of the object that needs to be updated as well as
-        a dictionary of k/v pairs used to update an object. The object
-        is directly updated on the server using a PATCH request without
-        fetching object information.
+        Allows for bulk updating of existing objects on an endpoint.
+        Objects is a list which contain either json/dicts or Record
+        derived objects, which contain the updates to apply.
+        If json/dicts are used, then the id of the object *must* be
+        included
 
-        For fields requiring an object reference (such as a device location),
-        the API user is responsible for providing the object ID or the object
-        URL. This API will not accept the pynautobot object directly.
+        :arg list,optional \*args: A list of dicts or a list of Record
 
-        :arg str id: Identifier of the object being updated
-        :arg dict data: Dictionary containing the k/v to update the
-            record object with.
-        :returns: True if PATCH request was successful.
+        :arg str,optional \**kwargs:
+            See Below
+
+        :Keyword Arguments:
+            * *id* (``string``) -- Identifier of the object being updated
+            * *data* (``dict``) -- k/v to update the record object with
+
+        :returns: A list or single :py:class:`.Record` object depending
+            on whether a bulk update was requested.
         :example:
 
+        Accepts the id of the object that needs to be updated as well as a
+            dictionary of k/v pairs used to update an object
         >>> nb.dcim.devices.update(id="0238a4e3-66f2-455a-831f-5f177215de0f", data={
-        ...     "name": "test-switch2",
-        ...     "serial": "ABC321",
+        ...     "name": "test",
+        ...     "serial": "1234",
         ...     "location": "9b1f53c7-89fa-4fb2-a89a-b97364fef50c",
         ... })
-        True
+        >>>
+
+        Use bulk update by passing a list of dicts:
+
+        >>> devices = nb.dcim.devices.update([
+        ...    {'id': "db8770c4-61e5-4999-8372-e7fa576a4f65", 'name': 'test'},
+        ...    {'id': "e9b5f2e0-4f20-41ad-9179-90a4987f743e", 'name': 'test2'},
+        ... ])
+        >>>
+
+        Use bulk update by passing a list of Records:
+
+        >>> devices = list(nb.dcim.devices.filter())
+        >>> devices
+        [Device1, Device2, Device3]
+        >>> for d in devices:
+        ...     d.name = d.name+'-test'
+        ...
+        >>> nb.dcim.devices.update(devices)
+        >>>
         """
+        if not args and not kwargs:
+            raise ValueError("You must provide either a UUID and data dict or a list of objects to update")
+        uuid = kwargs.get("id", "")
+        data = kwargs.get("data", {})
+        if data and not uuid:
+            uuid = args[0]
+        if len(args) == 2:
+            uuid, data = args
+
+        if not any([uuid, data]):
+            return self.bulk_update(args[0])
+
         req = Request(
-            key=id,
+            key=uuid,
             base=self.url,
             token=self.api.token,
             http_session=self.api.http_session,
@@ -345,6 +384,103 @@ class Endpoint(object):
         if req.patch(data):
             return True
         return False
+
+    def bulk_update(self, objects: List[Dict[str, Any]]):
+        r"""This method is called from the update() method if a bulk
+        update is detected.
+
+        Allows for bulk updating of existing objects on an endpoint.
+        Objects is a list which contain either json/dicts or Record
+        derived objects, which contain the updates to apply.
+        If json/dicts are used, then the id of the object *must* be
+        included
+
+        :arg list,optional \*args: A list of dicts or a list of Record
+        """
+        if not isinstance(objects, list):
+            raise ValueError("objects must be a list[dict()|Record] not " + str(type(objects)))
+
+        bulk_data = []
+        for o in objects:
+            try:
+                if isinstance(o, dict):
+                    bulk_data.append(o)
+                elif isinstance(o, Record):
+                    if not hasattr(o, "id"):
+                        raise ValueError("'Record' object has no attribute 'id'")
+                    updates = o.updates()
+                    if updates:
+                        updates["id"] = o.id
+                        bulk_data.append(updates)
+                else:
+                    raise ValueError("Invalid object type: " + str(type(o)))
+            except ValueError as exc:
+                raise ValueError("Unexpected value in object list") from exc
+
+        req = Request(
+            base=self.url,
+            token=self.api.token,
+            http_session=self.api.http_session,
+            api_version=self.api.api_version,
+        ).patch(bulk_data)
+        return response_loader(req, self.return_obj, self)
+
+    def delete(self, objects):
+        r"""Bulk deletes objects on an endpoint.
+
+        Allows for batch deletion of multiple objects from
+        a single endpoint
+
+        :arg list objects: A list of either ids or Records to delete.
+        :returns: True if bulk DELETE operation was successful.
+
+        :Examples:
+
+        Deleting all `devices`:
+
+        >>> pynautobot.dcim.devices.delete(pynautobot.dcim.devices.all())
+        >>>
+
+        Use bulk deletion by passing a list of ids:
+
+        >>> pynautobot.dcim.devices.delete(["db8770c4-61e5-4999-8372-e7fa576a4f65"
+        ...                                  ,"e9b5f2e0-4f20-41ad-9179-90a4987f743e"])
+        >>>
+
+        Use bulk deletion to delete objects eg. when filtering
+        on a `custom_field`:
+
+        >>> pynautobot.dcim.devices.delete([
+        ...         d for d in pynautobot.dcim.devices.all()
+        ...             if d.custom_fields.get("field", False)
+        ...     ])
+        >>>
+        """
+        ids = []
+        if not isinstance(objects, list):
+            raise ValueError("objects must be a list[str(id)|Record] not " + str(type(objects)))
+        for o in objects:
+            try:
+                if isinstance(o, str):
+                    if UUID(o):
+                        ids.append(o)
+                elif isinstance(o, Record):
+                    if not hasattr(o, "id"):
+                        raise ValueError("'Record' object has no attribute 'id'")
+                    ids.append(o.id)
+                else:
+                    raise ValueError("Invalid object type: " + str(type(o)))
+            except ValueError as exc:
+                raise ValueError("Unexpected value in object list") from exc
+
+        req = Request(
+            base=self.url,
+            token=self.token,
+            http_session=self.api.http_session,
+            api_version=self.api.api_version,
+        )
+
+        return req.delete(data=[{"id": id} for id in ids])
 
     def choices(self, api_version=None):
         """Returns all choices from the endpoint.
@@ -472,12 +608,16 @@ class DetailEndpoint(object):
         Returns the response from Nautobot for a detail endpoint.
 
         Args:
-            :arg str,optional api_version: Override default or globally set Nautobot REST API version for this single request.
-            **kwargs: key/value pairs that get converted into url parameters when passed to the endpoint.
-                E.g. ``.list(method='get_facts')`` would be converted to ``.../?method=get_facts``.
 
-        :returns: A dictionary or list of dictionaries retrieved from
-            Nautobot.
+        :arg str,optional api_version: Override default or globally set Nautobot REST API version for this single request.
+        :arg \**kwargs:
+            See below
+
+        :Keyword Arugments:
+            key/value pairs that get converted into url parameters when passed to the endpoint.
+            E.g. ``.list(method='get_facts')`` would be converted to ``.../?method=get_facts``.
+
+        :returns: A dictionary or list of dictionaries retrieved from Nautobot.
         """
         api_version = api_version or self.parent_obj.api.api_version
 
@@ -512,7 +652,7 @@ class DetailEndpoint(object):
 
 
 class RODetailEndpoint(DetailEndpoint):
-    def create(self, data):
+    def create(self, data=None, api_version=None):
         raise NotImplementedError("Writes are not supported for this endpoint.")