+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## v2.1.0
+
+### New Features
+
+- (#163) Adds `Endpoint.delete` method for bulk deleting of records
+- (#165) Adds `Endpoint.update` method for bulk updating of records
+
+### Fixes
+
+- (#162) Corrects signature of `RODetailEndpoint.create` to provide a proper error that it is not implemented when using `api_version`
+
 ## v2.0.2
 
 ### Fixes